requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
PyJWT==2.8.0